+++ src/include/libpmemobj++/detail/make_atomic_impl.hpp
@@ -39,6 +39,7 @@
 #define LIBPMEMOBJ_MAKE_ATOMIC_IMPL_HPP
 
 #include <new>
+#include <stddef.h>
 
 #include "libpmemobj++/detail/array_traits.hpp"
 #include "libpmemobj++/detail/integer_sequence.hpp"
