+++ src/include/libpmemobj++/make_persistent.hpp
@@ -40,7 +40,7 @@
 #ifndef PMEMOBJ_MAKE_PERSISTENT_HPP
 #define PMEMOBJ_MAKE_PERSISTENT_HPP
 
-#include "libpmemobj.h"
+#include "libpmemobj/tx_base.h"
 #include "libpmemobj++/detail/check_persistent_ptr_array.hpp"
 #include "libpmemobj++/detail/common.hpp"
 #include "libpmemobj++/detail/pexceptions.hpp"
