+++ src/include/libpmemobj++/make_persistent_array.hpp
@@ -40,7 +40,7 @@
 #ifndef PMEMOBJ_MAKE_PERSISTENT_ARRAY_HPP
 #define PMEMOBJ_MAKE_PERSISTENT_ARRAY_HPP
 
-#include "libpmemobj.h"
+#include "libpmemobj/tx_base.h"
 #include "libpmemobj++/detail/array_traits.hpp"
 #include "libpmemobj++/detail/check_persistent_ptr_array.hpp"
 #include "libpmemobj++/detail/common.hpp"
