+++ src/include/libpmemobj++/mutex.hpp
@@ -38,7 +38,8 @@
 #ifndef PMEMOBJ_MUTEX_HPP
 #define PMEMOBJ_MUTEX_HPP
 
-#include "libpmemobj.h"
+#include "libpmemobj/thread.h"
+#include "libpmemobj/tx_base.h"
 #include "libpmemobj++/detail/pexceptions.hpp"
 
 namespace nvml
