+++ src/include/libpmemobj++/pool.hpp
@@ -39,8 +39,9 @@
 #define PMEMOBJ_POOL_HPP
 
 #include <sys/stat.h>
+#include <stddef.h>
 
-#include "libpmemobj.h"
+#include "libpmemobj/pool_base.h"
 #include "libpmemobj++/detail/pexceptions.hpp"
 #include "libpmemobj++/p.hpp"
