+++ src/test/obj_include/.gitignore
@@ -1,0 +1,2 @@
+obj_inc
+*.tmpo
